requirements.txt
pandas==1.4.3
requests==2.28.1
jupyter==1.0.0
streamlit==1.11.1
python-dotenv==0.20.0
python-dotenv==0.20.0